PRDT-171 Improving authorizer for public vs administrative calls
