+++ lib/cdk/apigateway.js
@@ -8,14 +8,6 @@ const cdk = require('aws-cdk-lib');
 function apigwSetup(scope, props) {
   console.log('Setting up API Gateway resources...');
 
-  // TOKEN AUTHORIZER
-
-  const requestAuthorizer = new apigateway.RequestAuthorizer(scope, 'RequestAuthorizer', {
-    handler: props.authFunction,
-    identitySources: [apigateway.IdentitySource.header('Authorization')],
-    authorizerName: 'ReserveRecAuthorizer',
-  });
-
   // API GATEWAY
 
   const api = new apigateway.RestApi(scope, 'ApiDeployment', {
@@ -30,7 +22,7 @@ function apigwSetup(scope, props) {
       maxAge: cdk.Duration.seconds(600),
     },
     defaultMethodOptions: {
-      authorizer: requestAuthorizer,
+      authorizer: props.publicRequestAuthorizer,
     },
     deployOptions: {
       stageName: props.env.API_STAGE,
