+++ lib/handlers/users/GET/index.js
@@ -8,7 +8,7 @@ exports.handler = async function (event, context) {
 
   const userPoolId = event?.pathParameters?.userPoolId || null;
   let filter = null;
-  if (Object.keys(event?.queryStringParameters)?.length > 0) {
+  if (event?.queryStringParameters && Object.keys(event?.queryStringParameters)?.length > 0) {
     for (const key of LIST_USER_FILTER_KEYS) {
       if (event.queryStringParameters[key]) {
         // Cognito API can only filter by one key at a time. Other keys will be ignored.
