+++ lib/handlers/users/resources.js
@@ -28,19 +28,17 @@ function userSetup(scope, props) {
   });
   userPoolIdResource.addMethod(
     "GET",
-    new apigateway.LambdaIntegration(usersGetFunction)
-    // TODO: re-enable authorizer
-    // {
-    //   authorizer: props.authorizer,
-    // })
+    new apigateway.LambdaIntegration(usersGetFunction),
+    {
+      authorizer: props.adminRequestAuthorizer,
+    }
   );
   userSubResource.addMethod(
     "GET",
-    new apigateway.LambdaIntegration(usersGetFunction)
-    // TODO: re-enable authorizer
-    // {
-    //   authorizer: props.authorizer,
-    // })
+    new apigateway.LambdaIntegration(usersGetFunction),
+    {
+      authorizer: props.adminRequestAuthorizer,
+    }
   );
 
   // Add DynamoDB policy to the function
