+++ lib/reserve-rec-cdk-stack.js
@@ -67,7 +67,8 @@ class ReserveRecCdkStack extends Stack {
     // API GATEWAY
     const apigwResources = apigwSetup(this, {
       env: props.env,
-      authFunction: authorizerResources.authFunction,
+      publicRequestAuthorizer: authorizerResources.publicRequestAuthorizer,
+      adminRequestAuthorizer: authorizerResources.adminRequestAuthorizer,
     });
 
     // WEB SOCKET API
@@ -262,7 +263,7 @@ class ReserveRecCdkStack extends Stack {
         layerResources.awsUtilsLayer,
         layerResources.dataUtilsLayer
       ],
-      authorizer: authorizerResources.authFunction
+      adminRequestAuthorizer: authorizerResources.adminRequestAuthorizer
     });
 
     // WebSocket
@@ -278,19 +279,19 @@ class ReserveRecCdkStack extends Stack {
         ],
         pubsubTable: dynamodbResources.pubsubTable
       });
-    }
 
-    // Worldline Notification Webhook
-    const worldlineNotification = worldlineNotificationSetup(this, {
-      env: props.env,
-      api: apigwResources.api,
-      layers: [
-        layerResources.baseLayer,
-        layerResources.awsUtilsLayer,
-        layerResources.dataUtilsLayer
-      ],
-      mainTable: dynamodbResources.mainTable
-    });
+      // Worldline Notification Webhook
+      const worldlineNotification = worldlineNotificationSetup(this, {
+        env: props.env,
+        api: apigwResources.api,
+        layers: [
+          layerResources.baseLayer,
+          layerResources.awsUtilsLayer,
+          layerResources.dataUtilsLayer
+        ],
+        mainTable: dynamodbResources.mainTable
+      });
+    }
 
   }
 }
